+++ libythtbbs/binaryattach.h
@@ -1,6 +1,5 @@
 #ifndef __BINARYATTACH_H
 #define __BINARYATTACH_H
-#include <stdio.h>
 int appendbinaryattach(char *filename, char *userid, char *attachname);
-char *checkbinaryattach(char *buf, FILE * fp, unsigned int *len);
+char *checkbinaryattach(char *buf, FILE * fp, size_t *len);
 #endif
